Equipment.js
Refactored Equipment.js and EquipmentItem.js
