+++ src/helper/UserContext.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useContext, useEffect } from "react";
 import { DataStore } from "@aws-amplify/datastore";
 
-import { OrgUserStorage } from "../models";
+import { OrgUserStorage, UserOrStorage } from "../models";
 import { OrgType, UserType, UserContextType } from "../types/ContextTypes";
 
 /* Context that distributes the user object, organization object, and the user's organization object
@@ -25,6 +25,7 @@ export const UserProvider = ({ children }: { children: React.ReactNode }) => {
         c.and((c) => [
           c.organization.name.eq(org.name),
           c.user.userId.eq(user.attributes.sub),
+          c.type.eq(UserOrStorage.USER),
         ]),
       );
       setOrgUserStorage(orgUser[0]);
