+++ src/types/EquipmentTypes.ts
@@ -1,0 +1,6 @@
+export type EquipmentObj = {
+  id: string;
+  label: string;
+  count: number;
+  data: string[];
+};
